requirements.txt
fastapi~=0.115.8
uvicorn
sqlalchemy~=2.0.37
alembic
pydantic
requests~=2.32.3
python-dotenv~=1.0.1
loguru~=0.7.3

pydantic-settings~=2.7.1
pip~=25.0
distro~=1.9.0
wheel~=0.38.1